+++ sync/sync-impl/src/main/java/com/duckduckgo/sync/impl/ui/setup/SetupAccountActivity.kt
@@ -16,13 +16,10 @@
 
 package com.duckduckgo.sync.impl.ui.setup
 
-import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.widget.Toast
 import androidx.activity.OnBackPressedCallback
-import androidx.activity.result.contract.ActivityResultContract
 import androidx.fragment.app.commitNow
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.flowWithLifecycle
@@ -34,7 +31,6 @@ import com.duckduckgo.mobile.android.ui.viewbinding.viewBinding
 import com.duckduckgo.sync.impl.R.id
 import com.duckduckgo.sync.impl.databinding.ActivitySyncSetupAccountBinding
 import com.duckduckgo.sync.impl.ui.SyncDeviceConnectedFragment
-import com.duckduckgo.sync.impl.ui.SyncLoginActivity
 import com.duckduckgo.sync.impl.ui.setup.SetupAccountActivity.Companion.Screen.DEVICE_CONNECTED
 import com.duckduckgo.sync.impl.ui.setup.SetupAccountActivity.Companion.Screen.RECOVERY_CODE
 import com.duckduckgo.sync.impl.ui.setup.SetupAccountActivity.Companion.Screen.SETUP
@@ -46,28 +42,10 @@ import com.duckduckgo.sync.impl.ui.setup.SetupAccountViewModel.ViewMode.AskSyncA
 import com.duckduckgo.sync.impl.ui.setup.SetupAccountViewModel.ViewMode.DeviceConnected
 import com.duckduckgo.sync.impl.ui.setup.SetupAccountViewModel.ViewMode.TurnOnSync
 import com.duckduckgo.sync.impl.ui.setup.SetupAccountViewModel.ViewState
-import com.duckduckgo.sync.impl.ui.setup.SyncSetupFlowFragment.SetupFlowListener
 import com.duckduckgo.sync.impl.ui.setup.SyncSetupFlowViewModel.ViewMode.InitialSetupScreen
 import com.duckduckgo.sync.impl.ui.setup.SyncSetupFlowViewModel.ViewMode.SyncAnotherDeviceScreen
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import timber.log.Timber
-
-class LoginContract : ActivityResultContract<Void?, Boolean>() {
-    override fun createIntent(
-        context: Context,
-        input: Void?,
-    ): Intent {
-        return SyncLoginActivity.intent(context)
-    }
-
-    override fun parseResult(
-        resultCode: Int,
-        intent: Intent?,
-    ): Boolean {
-        return resultCode == Activity.RESULT_OK
-    }
-}
 
 @InjectWith(ActivityScope::class)
 class SetupAccountActivity : DuckDuckGoActivity(), SetupFlowListener {
@@ -78,7 +56,6 @@ class SetupAccountActivity : DuckDuckGoActivity(), SetupFlowListener {
 
     private val loginFlow = registerForActivityResult(LoginContract()) { resultOk ->
         if (resultOk) {
-            Toast.makeText(this, "login went well", Toast.LENGTH_LONG).show()
             viewModel.onLoginSucess()
         }
     }
@@ -90,7 +67,6 @@ class SetupAccountActivity : DuckDuckGoActivity(), SetupFlowListener {
         } else {
             screen = intent.getSerializableExtra(SETUP_ACCOUNT_SCREEN_EXTRA) as? Screen ?: SETUP
         }
-        Timber.i("CRIS: onCreate $screen")
         setContentView(binding.root)
         observeUiEvents()
         configureListeners()
