+++ sync/sync-impl/src/main/res/layout/fragment_device_connected.xml
